+++ pymodbus/pdu/file_message.py
@@ -63,7 +63,8 @@ class ReadFileRecordRequest(ModbusPDU):
     def decode(self, data: bytes) -> None:
         """Decode the incoming request."""
         self.records = []
-        byte_count = int(data[0])
+        if (byte_count := int(data[0])) > len(data) - 1:
+            raise ModbusException(f"Invalid byte count: {byte_count}")
         for count in range(1, byte_count, 7):
             decoded = struct.unpack(">BHHH", data[count : count + 7])
             record = FileRecord(
@@ -169,7 +170,8 @@ class WriteFileRecordRequest(ModbusPDU):
 
     def decode(self, data: bytes) -> None:
         """Decode the incoming request."""
-        byte_count = int(data[0])
+        if (byte_count := int(data[0])) > len(data) - 1:
+            raise ModbusException(f"Invalid byte count: {byte_count}")
         count = 1
         self.records.clear()
         while count < byte_count:
@@ -293,7 +295,7 @@ class ReadFifoQueueResponse(ModbusPDU):
         """Calculate the size of the message."""
         hi_byte = int(data[2])
         lo_byte = int(data[3])
-        return ((hi_byte << 16) + lo_byte) + 4
+        return ((hi_byte << 8) + lo_byte) + 4
 
     def __init__(
         self, values: list[int] | None = None, dev_id: int = 1, transaction_id: int = 0
@@ -314,6 +316,8 @@ class ReadFifoQueueResponse(ModbusPDU):
         """Decode a the response."""
         self.values = []
         _, count = struct.unpack(">HH", data[0:4])
+        if 4 + count * 2 > len(data):
+            raise ModbusException(f"Invalid fifo count: {count}")
         for index in range(0, count):
             idx = 4 + index * 2
             self.values.append(struct.unpack(">H", data[idx : idx + 2])[0])