+++ pymodbus/simulator/simdevice.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import inspect
-from collections.abc import Awaitable, Callable
+from collections.abc import Callable, Coroutine
 from dataclasses import dataclass
 from typing import TypeAlias, cast
 
@@ -15,7 +15,7 @@ from .simutils import DataType, SimUtils
 
 SimAction: TypeAlias = Callable[
     [int, int, int, int, list[int], list[int] | list[bool] | None],
-    Awaitable[None | ExcCodes],
+    Coroutine[None, None, ExcCodes | None],
 ]
 SimRegs: TypeAlias = tuple[int, list[int], list[int]]
 TUPLE_NAMES = ("coils", "discrete inputs", "holding registers", "input registers")
@@ -101,12 +101,15 @@ class SimDevice:
     #:         count: int,            # request count
     #:         current_registers: list[int],  # current registers (modify inline)
     #:         set_values: list[int] | list[bool] | None  # request values to be written (None for read requests)
-    #      ) -> None | ExceptionResponse:
+    #:     ) -> ExcCodes | None:
     #:
     #: action can:
+    #:
     #: - update registers (affect the current and future responses)
+    #:
     #: - update set_values (affect the register update)
-    #: - return an ExceptionResponse.
+    #:
+    #: - return a member of :class:`pymodbus.constants.ExcCodes`.
     #:
     #: .. tip:: use functools.partial to add extra parameters if needed.
     action: SimAction | None = None
@@ -149,7 +152,7 @@ class SimDevice:
                 "simdata= must be SimData, list of SimData or tuple with 4 list of SimData"
             )
         for i in range(4):
-            sim_list = cast(tuple, self.simdata)[i]
+            sim_list = self.simdata[i]
             if not isinstance(sim_list, list):
                 raise TypeError(f"simdata=tuple[{TUPLE_NAMES[i]}] -> must be a list")
             for inx, entry in enumerate(sim_list):
@@ -186,9 +189,7 @@ class SimDevice:
         self.__check_simple2()
         if isinstance(self.simdata, tuple):
             for i in range(4):
-                self.__check_block(
-                    cast(tuple, self.simdata)[i], (i in {0, 1}), TUPLE_NAMES[i]
-                )
+                self.__check_block(self.simdata[i], (i in {0, 1}), TUPLE_NAMES[i])
         else:
             x_simdata = (
                 self.simdata if isinstance(self.simdata, list) else [self.simdata]
@@ -279,7 +280,7 @@ class SimDevice:
         #  (<coils>, <discrete inputs>, <holding registers>, <input registers>)
         convert = {0: "c", 1: "d", 2: "h", 3: "i"}
         for i in range(4):
-            x_simdata = cast(tuple, self.simdata)[i]
+            x_simdata = self.simdata[i]
             x_simdata.sort(key=lambda x: x.address)
             if i in {0, 1}:
                 b[convert[i]] = self.__create_block_bits(x_simdata)