+++ test/server/test_base.py
@@ -5,6 +5,7 @@ from unittest import mock
 import pytest
 
 from pymodbus.datastore import ModbusDeviceContext, ModbusServerContext
+from pymodbus.exceptions import ParameterException
 from pymodbus.framer import FramerType
 from pymodbus.pdu import ReadHoldingRegistersRequest
 from pymodbus.server import ModbusBaseServer
@@ -140,3 +141,27 @@ class TestBaseServer:
         """Test serve_forever."""
         with pytest.raises(RuntimeError):
             baseserver.callback_data(None)
+
+    async def test_base_old_context(self):
+        """Test serve_forever."""
+        context = ModbusServerContext(devices=ModbusDeviceContext())
+        context.simdevices = []
+        with pytest.raises(ParameterException):
+            ModbusBaseServer(
+                CommParams(
+                    comm_type=CommType.TCP,
+                    comm_name="server_listener",
+                    reconnect_delay=0.0,
+                    reconnect_delay_max=0.0,
+                    timeout_connect=0.0,
+                ),
+                context,
+                False,
+                False,
+                None,
+                FramerType.SOCKET,
+                None,
+                None,
+                None,
+                [ReadHoldingRegistersRequest],
+            )